[.NET] Move source generator into a pluggable python module
