+++ interfaces/sourcegen/csharp/__init__.py
@@ -1,0 +1,1 @@
+from ._sourcegen import SourceGenerator
